Analysis of the educational level of the school
It receives the Excel file and checks whether this file exists in the system and then checks again whether the column in the Excel file exists or not. 
And it sorts the grades from the lowest to the highest 
It removes empty cells and grades greater than 0 and greater than 20 and does not include them in the calculation and analyzes and analyzes the educational level by displaying a graph to announce.